Skip image search if there are no images on a page
When looking for the first image on a Wikipedia page with no images, `getUrlOfMainImage` throws an error `index 1 beyond bounds...`. presumably because there aren't any images in the newly-created array of images from the page's parsed HTML.

This just wraps the whole thing in a check to make sure there's something in the array before looping through it.